separated PrintingConfig class to Printer and PrintingConfig
Updates ObjectExtensions methods

Add new method to MemberInfoExtensions
